+++ packages/sdk/src/services/graphql-api.ts
@@ -9,6 +9,7 @@ import {
   ChatMessage,
   Status,
   ChatMessageReportedBy,
+  PublicUserInput,
 } from '@arena-im/chat-types';
 import { GraphQLTransport } from './graphql-transport';
 import { DEFAULT_AUTH_TOKEN } from '../config';
@@ -56,6 +57,8 @@ export class GraphQLAPI {
               image
               isBlocked
               status
+              isModerator
+              modLabel
             }
             name
             unreadCount
@@ -139,6 +142,8 @@ export class GraphQLAPI {
             image
             isBlocked
             status
+            isModerator
+            modLabel
           }
           name
           amIBlocked
@@ -169,6 +174,7 @@ export class GraphQLAPI {
               status
               modLabel
               isModerator
+              modLabel
               image
               isBlocked
             }
@@ -214,6 +220,8 @@ export class GraphQLAPI {
             image
             isBlocked
             status
+            isModerator
+            modLabel
           }
           name
           unreadCount
@@ -665,5 +673,101 @@ export class GraphQLAPI {
     }
 
     return channel;
+  }
+
+  public async fetchUserProfile(userId: string): Promise<PublicUser> {
+    const query = gql`
+      query user($id: ID!) {
+        user(id: $id) {
+          _id
+          name
+          image
+          defaultImage
+          isModerator
+          bio
+          socialLinks {
+            url
+            provider
+          }
+          isBlocked
+          userName
+          location
+          slug
+        }
+      }
+    `;
+
+    const data = await this.graphQL.client.request(query, { id: userId });
+
+    const user = data.user as PublicUser;
+
+    if (!user) {
+      throw new Error(Status.Invalid);
+    }
+
+    return user;
+  }
+
+  public async updateUser(user: PublicUserInput): Promise<PublicUser> {
+    const mutation = gql`
+      mutation updateUser($input: PublicUserInput!) {
+        updateUser(input: $input) {
+          _id
+          name
+          image
+          defaultImage
+          isModerator
+          bio
+          socialLinks {
+            url
+            provider
+          }
+          isBlocked
+          userName
+          location
+          slug
+        }
+      }
+    `;
+
+    const input: PublicUserInput = {};
+
+    if (typeof user.bio !== 'undefined') {
+      input.bio = user.bio;
+      input.socialLinks = user.socialLinks;
+    }
+    if (typeof user.location !== 'undefined') {
+      input.location = user.location;
+    }
+
+    if (typeof user.name !== 'undefined') {
+      input.name = user.name;
+    }
+
+    if (typeof user.socialLinks !== 'undefined') {
+      input.socialLinks = user.socialLinks;
+    }
+
+    if (typeof user.image !== 'undefined') {
+      input.image = user.image;
+    }
+
+    if (typeof user.useDefaultImage !== 'undefined') {
+      input.useDefaultImage = user.useDefaultImage;
+    }
+
+    if (typeof user.password !== 'undefined') {
+      input.password = user.password;
+    }
+
+    const data = await this.graphQL.client.request(mutation, { input });
+
+    const result = data.updateUser as PublicUser;
+
+    if (!result) {
+      throw new Error(Status.Failed);
+    }
+
+    return result;
   }
 }