+++ packages/types/src/index.ts
@@ -14,10 +14,20 @@ export { ChatRoom, ChatModerationRequest, LiveChatChannel, BaseLiveChat, BaseCha
 export { DocumentChangeType, ListenChangeConfig, Where, OrderBy } from './firestore';
 export { Site, EmbedSettings } from './site';
 export { Status } from './status';
-export { PublicUser, User, ExternalUser, BanUser, ProviderUser, UserChangedListener, SSOExchangeResult } from './user';
+export {
+  PublicUser,
+  User,
+  ExternalUser,
+  BanUser,
+  ProviderUser,
+  UserChangedListener,
+  SSOExchangeResult,
+  PublicUserInput,
+} from './user';
 export { Moderation, ModeratorStatus } from './moderation';
 export { PublicUserStatus, GroupChannel } from './private-chat';
 export { BasePrivateChannel } from './base-private-channel';
 export { BaseQna, QnaQuestion, QnaQuestionFilter, QnaProps } from './qna';
 export { BasePolls, Poll, PollFilter } from './polls';
 export { TrackPayload, TrackContext, TrackPageInfo } from './arena-hub';
+export { BaseUserProfile } from './user-profile';